chore: delete now unused code
