+++ wire-ios-request-strategy/Sources/Request Strategies/Generic Message/GenericMessageEntity.swift
@@ -136,59 +136,3 @@ extension GenericMessageEntity: EncryptedPayloadGenerator {
     }
 
 }
-
-/// This should not be used as a standalone strategy but either subclassed or used within another
-/// strategy. Please have a look at `CallingRequestStrategy` and `GenericMessageNotificationRequestStrategy`
-/// before modifying the behaviour of this class.
-@objcMembers public class GenericMessageRequestStrategy: OTREntityTranscoder<GenericMessageEntity>, ZMRequestGenerator, ZMContextChangeTracker {
-
-    private var sync: DependencyEntitySync<GenericMessageRequestStrategy>?
-    private var requestFactory = ClientMessageRequestFactory()
-
-    public override init(context: NSManagedObjectContext, clientRegistrationDelegate: ClientRegistrationDelegate) {
-        super.init(context: context, clientRegistrationDelegate: clientRegistrationDelegate)
-
-        sync = DependencyEntitySync(transcoder: self, context: context)
-    }
-
-    public func schedule(message: GenericMessage, inConversation conversation: ZMConversation, targetRecipients: GenericMessageEntity.Recipients = .conversationParticipants, completionHandler: ((_ response: ZMTransportResponse) -> Void)?) {
-        sync?.synchronize(entity: GenericMessageEntity(conversation: conversation, message: message, targetRecipients: targetRecipients, completionHandler: completionHandler))
-        RequestAvailableNotification.notifyNewRequestsAvailable(nil)
-    }
-
-    public func expireEntities(withDependency dependency: AnyObject) {
-        guard let dependency = dependency as? NSManagedObject else { return }
-        sync?.expireEntities(withDependency: dependency)
-    }
-
-    public override func request(forEntity entity: GenericMessageEntity, apiVersion: APIVersion) -> ZMTransportRequest? {
-        return requestFactory.upstreamRequestForMessage(entity, in: entity.conversation!, apiVersion: apiVersion)
-    }
-
-    public override func shouldTryToResend(entity: GenericMessageEntity, afterFailureWithResponse response: ZMTransportResponse) -> Bool {
-        entity.completionHandler?(response)
-        return super.shouldTryToResend(entity: entity, afterFailureWithResponse: response)
-    }
-
-    public override func request(forEntity entity: GenericMessageEntity, didCompleteWithResponse response: ZMTransportResponse) {
-        super.request(forEntity: entity, didCompleteWithResponse: response)
-
-        entity.completionHandler?(response)
-    }
-
-    public func nextRequest(for apiVersion: APIVersion) -> ZMTransportRequest? {
-        return sync?.nextRequest(for: apiVersion)
-    }
-
-    public func objectsDidChange(_ object: Set<NSManagedObject>) {
-        sync?.objectsDidChange(object)
-    }
-
-    public func fetchRequestForTrackedObjects() -> NSFetchRequest<NSFetchRequestResult>? {
-        return sync?.fetchRequestForTrackedObjects()
-    }
-
-    public func addTrackedObjects(_ objects: Set<NSManagedObject>) {
-        sync?.addTrackedObjects(objects)
-    }
-}
